==> 01 Select parts of typical Text Analytics pipeline.ipynb <==
# Text Analytics - Discovery (EDA)

import nltk

#These only need to be done once...
#nltk.download('punkt')                      
#nltk.download('averaged_perceptron_tagger') 


print((nltk.__version__))

## Tokenize into sentences and words

#one big text... 

text = u"""
On this day, we gather because we have chosen hope over fear, unity of purpose over conflict and discord. 
On this day, we come to proclaim an end to the petty grievances and false promises, 
the recriminations and worn-out dogmas that for far too long have strangled our politics. We remain a young 
nation. But in the words of Scripture, the time has come to set aside childish things. The time has 
come to reaffirm our enduring spirit; to choose our better history; to carry forward that precious gift, 
that noble idea passed on from generation to generation: the God-given promise that all are equal, all 
are free, and all deserve a chance to pursue their full measure of happiness.
"""

# Sentence Tokenization
sentnc = nltk.sent_tokenize(text)

# Word Tokenization
wrds = [nltk.word_tokenize(s) for s in sentnc]

print("Sentences: {:>5} \n    Words: {:>5}".format(len(sentnc), sum(len(x)for x in wrds)))

#TODO: dont forget to ask... Should we clean these before we use them?
sentnc[2]

wrds[2]

## Part of Speech (POS) Tagging

#This will create a list of (token, POS tag) for each token in the doc

tagged_wt = [nltk.pos_tag(w)for w in wrds]     #<<< using the word list

## [[('On', 'IN'), ('this', 'DT'), ('day', 'NN'), (',', ','), ('we', 'PRP') ...

patternPOS= []    

#add all the parts of speech into that list: patternPOS
for tag in tagged_wt:
    patternPOS.append([v for k,v in tag])


print(patternPOS)


## Extracting Nouns or noun like words

nouns = []       #what will I do with this empty list?

for tag in tagged_wt:
    nouns.append([k for k,v in tag if v in ['NN','NNS','NNP','NNPS']])
    
nouns[0]

verbs = [] 

for tag in tagged_wt:
    verbs.append([k for k,v in tag if v in ['VB','VBD','VBG','VBN','VBP','VBZ']])

verbs[0:2]

print("There are {} nouns and {} verbs in the text.".format(sum(len(l) for l in nouns), sum(len(l) for l in verbs)))

# spaCy

##### More feature extraction and explorations

spacy.__en_core_web_lg__

English multi-task CNN trained on OntoNotes, with GloVe vectors trained on Common Crawl. Assigns word vectors, context-specific token vectors, POS tags, dependency parse and named entities. (https://spacy.io/models)



#These need to be downloaded before we can use them. 
#!python -m spacy download en_pytt_xlnetbasecased_lg  
#!python -m spacy download en_core_web_lg

#python -m spacy download en  #<<-- get the English language models
import spacy

print((spacy.__version__))

import spacy
#the 'nlp' is the model
nlp = spacy.load('en_core_web_lg')  #the _lg (large) has the word embeddings

doc = nlp(text)

print(doc.text)

for token in doc:
    print(token.text, token.pos_, token.dep_)

spacy.explain("PROPN")

spacy.explain("ADP") #am I a linguist? What is this?

##  https://glossary.sil.org/term/adposition  

from collections import Counter
from tabulate import tabulate

#token.lemma_ is the base form of the token, with no inflectional suffixes.

noun_counter = Counter(token.lemma_ for token in doc if token.pos_ == 'NOUN')

print(tabulate(noun_counter.most_common(5), headers=['Noun', 'Count']))

print(noun_counter.most_common())

## Named Entity Recognition (NER)

doc = nlp(u"My name is Srinivas and I live in India.")

entity_types = ((ent.text, ent.label_) for ent in doc.ents)

print(tabulate(entity_types, headers=['Entity', 'Entity Type']))

NLTK POS Tagger meanings 
- __CC__ coordinating conjunction

- __CD__ cardinal digit

- __DT__ determiner

- __EX__ existential there (like: “there is” … think of it like “there exists”)

- __FW__ foreign word

- __IN__ preposition/subordinating conjunction

- __JJ__ adjective ‘big’

- __JJR__ adjective, comparative ‘bigger’

- __JJS__ adjective, superlative ‘biggest’

- __LS__ list marker 1)

- __MD__ modal could, will

- __NN__ noun, singular ‘desk’

- __NNS__ noun plural ‘desks’

- __NNP__ proper noun, singular ‘Harrison’

- __NNPS__ proper noun, plural ‘Americans’

- __PDT__ predeterminer ‘all the kids’

- __POS__ possessive ending parent‘s

- __PRP__ personal pronoun I, he, she

- __PRP$__ possessive pronoun my, his, hers

- __RB__ adverb very, silently,

- __RBR__ adverb, comparative better

- __RBS__ adverb, superlative best

- __RP__ particle give up

- __TO__ to go ‘to‘ the store.

- __UH__ interjection errrrrrrrm

- __VB__ verb, base form take

- __VBD__ verb, past tense took

- __VBG__ verb, gerund/present participle taking

- __VBN__ verb, past participle taken

- __VBP__ verb, sing. present, non-3d take

- __VBZ__ verb, 3rd person sing. present takes

- __WDT__ wh-determiner which

- __WP__ wh-pronoun who, what

- __WP$__ possessive wh-pronoun whose

- __WRB__ wh-abverb where, when

__References:__

https://www.nltk.org/book/

http://www.nltk.org/howto/

https://www.geeksforgeeks.org/part-speech-tagging-stop-words-using-nltk-python/  

https://spacy.io/api

Bird, Steven, Edward Loper and Ewan Klein (2009), Natural Language Processing with Python. O’Reilly Media Inc.

Tukey, John (1977), Exploratory Data Analysis, Addison-Wesley.

Brown, D. W. (2016) Corpus of Presidential Speeches. Retrieved from http://www.thegrammarlab.com
