==> 02 - Load all text files and pickle.ipynb <==
# Presidential Speech similarity

Data: 
Brown, D. W. (2016) Corpus of Presidential Speeches. Retrieved from http://www.thegrammarlab.com

Data link: http://www.thegrammarlab.com/?nor-portfolio=corpus-of-presidential-speeches-cops-and-a-clintontrump-corpus


import pandas as pd
import os
import pickle

root_dir = os.getcwd()
data_path = os.path.join(root_dir, "Data\\Corpus of Presential Speeches")

#these are all the source directories
data_dir = os.listdir(data_path)
print(data_dir)

import re

title_regex = re.compile("<title=\"(.+?)\">")

date_regex = re.compile("<date=\"(.+?)\">")



def file_to_df(file_path, president, filename):

    text = open(file_path, 'r', encoding="utf8")
    
    contents =text.read()
    
    title = re.findall(title_regex, str(contents))
    
    pub_date = re.findall(date_regex, contents)
    
    contents = re.sub(title_regex, '', contents)
    contents = re.sub(date_regex, '', contents)
    contents = re.sub(r'^\n\n','',contents)
    
    df =  pd.DataFrame({'filename': filename, 'president': president, 'title' : title, 'pub_date': pub_date, 'speech': contents})

    return df
    


df = pd.DataFrame()

for d in data_dir:
    for j,k,f in os.walk((os.path.join(data_path, d))):
        for s in f:
            df = df.append( file_to_df( (os.path.join(data_path, d, s)), d, s) )

df.reset_index(inplace=True, drop=True)
df.head()


root_dir = os.getcwd()
output_path = os.path.join(root_dir, "Output")

file_name = os.path.join(output_path,"president_speeches.pkl")

df.to_pickle(file_name)
