==> 03 Text analysis - spaCy.ipynb <==
# Presidential Speech similarity (cont)

from datetime import datetime
import pandas as pd
import pickle
import os

from collections import Counter
from tabulate import tabulate


#!python -m spacy download en_core_web_sm
#!python -m spacy download en_core_web_lg

#Load our pickle
root_dir = os.getcwd()

output_path = os.path.join(root_dir, "Output")

file_name = os.path.join(output_path,"president_speeches.pkl")

df = pd.read_pickle(file_name)

df.head()

df.title

import string
import spacy

from spacy.lang.en.stop_words import STOP_WORDS
from spacy.lang.en import English

nlp = spacy.load('en_core_web_lg')  #<< switch to the 'large' model with embeddings


## Digging deeper on named entities NER

While it is possible to make/add custom entity lists, we'll not do that today.

https://towardsdatascience.com/custom-named-entity-recognition-using-spacy-7140ebbb3718


#pick a speech and look at some entities ()
print(df.pub_date[5])

speech = nlp(df.speech[5])

entities=[(token, token.label_, token.label) for token in speech.ents]  

entities[0:5]


spacy.explain("GPE")  #aka geo-political entities

from spacy import displacy

#spaCy.displacy helps visualize embeddings and such
displacy.render(speech, style = "ent", jupyter = True)

#select one speech (Presidend Adams Inagural Speech)
adams_inaug = df[ (df['president'] == 'adams') & (df['title'] == 'Inaugural Address') ]["speech"].values[0]

adams_inaug

#pick a speech
adams_inaug = df[ (df['president'] == 'adams') & (df['title'] == 'Inaugural Address') ]["speech"].values[0]

#nlp process the speech
speech = nlp(adams_inaug)


#what type of entities are in this speech?
entity_types = ((ent.text, ent.label_) for ent in speech.ents)

print(tabulate(entity_types, headers=['Entity', 'Entity Type']))

#### Question

Are there any tokens/words that are not part of the spacy model vocabulary from that speech?

outs = []          #OOV is 'out of vocabulary'

for token in speech:

    if  token.is_oov:
        outs.append(token)    

print(outs)


#### Question

How can we use all those named entities?  

How can we use differnet parts of speech?

#                            LEMMA                             POS = NOUN
noun_counter = Counter(token.lemma_ for token in speech if token.pos_ == 'NOUN')

print(tabulate(noun_counter.most_common(10), headers=['Noun Lemmas', 'Count']))


#### Question
How many words in the spacy model embeddings vector?

print("There are {:,} word embeddings in this model.".format((speech.vocab.vectors.size)))

print(spacy.__version__)

### JSON 

#nlp keeps a lot of info on the document and can output that in json (easier to read)
import json

json_s = speech.to_json()   #not available in spacy < 2.2

with open("SpeechData.json", "w") as speech_data_file:
    json.dump(json_s, speech_data_file, indent=4) 
    


#### Question
How many sentences are in this speech?

print("There are {} senteneces in this speech".format(len(list(speech.sents))))


#### Question
How many Part of Speech (POS) tags are in this speech?  What are they?

speech_tags = []

for s in speech.sents:
    for token in s:
        speech_tags.append(str(token.pos_))

        
tag_counter = Counter(t for t in speech_tags)

print(tabulate(tag_counter.most_common(15), headers=['POS (token.pos_)','Count']))

#### Question
Spacy has several POS attributes.  What is the differnce between the __.pos___ and __.tag___?

#Tabulate parts of speech

speech_tags = []

for s in speech.sents:
    for token in s:
        speech_tags.append(str(token.tag_))

        
tag_counter = Counter(t for t in speech_tags)

print(tabulate(tag_counter.most_common(15), headers=['POS (token.tag_)','Count']))

#### Question
How many verbs are in this speech?

from collections import Counter
from tabulate import tabulate

verb_counter = Counter(token.lemma_ for token in speech if token.pos_ == 'VERB')

print(tabulate(verb_counter.most_common(5), headers=['Verbs', 'Count']))

# Semantic Textual Similarity


The default estimate is cosine similarity using an average of word vectors. [spaCy]

#Just a quick re-cap on how to compare two documents
doc  = nlp("Here is some text to encode.")
doc2 = nlp("Here are some words to encode.")
doc3 = nlp("Encoding words is what this does!")
doc4 = nlp("This workshop is boring because there is no MATLAB code!")
doc5 = nlp("My car needs gas.")
doc6 = nlp("Space is the final frontier.")

print(doc.similarity(doc2))
print(doc.similarity(doc3))
print(doc2.similarity(doc3))
print(doc.similarity(doc4))
print(doc.similarity(doc5))
print(doc.similarity(doc6))



__References:__

SCIBERT: A Pretrained Language Model for Scientific Text
https://arxiv.org/pdf/1903.10676.pdf

Text Similarity Estimation Based on Word Embeddings and Matrix
Norms for Targeted Marketing https://www.aclweb.org/anthology/N19-1181.pdf
