==> 04 End to End Similarity.ipynb <==
from datetime import datetime
import pandas as pd
import pickle
import os

from collections import Counter
from tabulate import tabulate

import spacy

from spacy.lang.en import English

nlp = spacy.load('en_core_web_lg')


#!python -m spacy download en_core_web_lg

root_dir = os.getcwd()
output_path = os.path.join(root_dir, "Output")

file_name = os.path.join(output_path,"president_speeches.pkl")

df = pd.read_pickle(file_name)
df.head()

df.describe()

print("There are {:,} possible combinations from {} speechs".format( (len(df) * len(df) ), (len(df))))

pres_counter = Counter(df.president)

print(tabulate(pres_counter.most_common(20), headers=['Pres', 'Speech Count']))

#who has 'Inaugural'
title_counter = Counter(df.title)

print(tabulate(title_counter.most_common(10), headers=['Title', 'Speech Count']))

df[df.title == 'Inaugural Address']

bush_inaug = df[ (df['president'] == 'bush') & (df['title'] == 'Inaugural Address') ]["speech"].values[0]

obama_inaug = df[ (df['president'] == 'obama') & (df['title'] == 'Inaugural Address') ]["speech"].values[0]

vanburen_inaug = df[ (df['president'] == 'vanburen') & (df['title'] == 'Inaugural Address') ]["speech"].values[0]


#Get a document that has been through the spacy."NLP" process
bush_doc = nlp(bush_inaug)

obama_doc = nlp(obama_inaug)

vanburen_doc = nlp(vanburen_inaug)

#### Question
What is the semantic textual similarity between two documents?

bush_doc.similarity(obama_doc)

bush_doc.similarity(vanburen_doc)

df[df.title == 'Address to the Nation About the Watergate Investigations']

nixon_before = df[ (df['filename'] == 'nixon_speeches_017.txt') ]["speech"].values[0] 
nixon_after = df[ (df['filename'] == 'nixon_speeches_016.txt') ]["speech"].values[0]

nixon_before_doc = nlp(nixon_before)
nixon_after_doc = nlp(nixon_after)

nixon_after_doc.similarity(nixon_before_doc)

nixon_after_doc.similarity(vanburen_doc)

#### Question
Where do we go from here?

Should we compare every sentence from both speeches?
